Moved org.bukkit.Vector to org.bukkit.util.
